refactor: 컨설턴트 등록 API 수정
