--- src/main/java/site/cleanfree/be_main/consultant/ConsultantRegisterRequestVo.java
@@ -1,9 +1,0 @@
-package site.cleanfree.be_main.consultant;
-
-import lombok.Getter;
-
-@Getter
-public class ConsultantRegisterRequestVo {
-
-    private String ip;
-}
