+++ src/main/java/site/cleanfree/be_main/consultant/vo/ConsultantRegisterRequestVo.java
@@ -1,0 +1,5 @@
+package site.cleanfree.be_main.consultant.vo;
+
+public record ConsultantRegisterRequestVo(String name, String contact) {
+
+}
